requirements.txt
Django==4.2.1
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2==2.9.9